+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 - Automatic check for missing dependencies of assembly addons on *assembly sync*.
+- Ability to use existing assembly for project via `odood assembly use` command.
+  This could be useful in CI to automate assembly sync process.
 
 ### Changed
 - `odood log` now will automatically show the end of logfile